% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflendzoneApp-package.R
\docType{package}
\name{nflendzoneApp-package}
\alias{nflendzoneApp}
\alias{nflendzoneApp-package}
\title{nflendzoneApp: NFL EndZone Analytics Shiny Dashboard}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Interactive Shiny dashboard for NFL team rankings, historical performance and Bayesian betting predictions (built with nflverse & cmdstanr).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TylerPollard410/nflendzoneApp}
  \item Report bugs at \url{https://github.com/TylerPollard410/nflendzoneApp/issues}
}

}
\author{
\strong{Maintainer}: Tyler Pollard \email{tylerpollard410@gmail.com}

}
\keyword{internal}
